Use common DSL for existing data formats and jackson-based ones
