+++ components/camel-jackson-protobuf/src/main/java/org/apache/camel/component/jackson/protobuf/JacksonProtobufDataFormat.java
@@ -26,7 +26,7 @@ import org.apache.camel.spi.annotations.Dataformat;
  * Marshal POJOs to Protobuf and back using <a href="http://jackson.codehaus.org/">Jackson</a>
  */
 @Dataformat("protobuf-jackson")
-@Metadata(excludeProperties = "library,permissions,dropRootNode")
+@Metadata(firstVersion = "3.10.0", excludeProperties = "contentTypeFormat,defaultInstance")
 public class JacksonProtobufDataFormat extends AbstractJacksonDataFormat {
 
     /**
