+++ core/camel-core-reifier/src/main/java/org/apache/camel/reifier/dataformat/JsonDataFormatReifier.java
@@ -85,7 +85,7 @@ public class JsonDataFormatReifier extends DataFormatReifier<JsonDataFormat> {
         }
         if (definition.getLibrary() == JsonLibrary.Jackson) {
             properties.put("schemaResolver", asRef(definition.getSchemaResolver()));
-            properties.put("autoDiscoverSchemaResolver", definition.isAutoDiscoverSchemaResolver());
+            properties.put("autoDiscoverSchemaResolver", definition.getAutoDiscoverSchemaResolver());
         }
     }
