+++ core/camel-util/src/main/java/org/apache/camel/util/SensitiveUtils.java
@@ -30,18 +30,13 @@ public final class SensitiveUtils {
                     "accesskey",
                     "accesstoken",
                     "accesstokensecret",
-                    "accountkey",
                     "accountsid",
                     "acltoken",
                     "authorizationtoken",
-                    "blobaccesskey",
-                    "blobstoragesharedkeycredential",
                     "certresourcepassword",
                     "clientsecret",
-                    "connectionstring",
                     "consumerkey",
                     "consumersecret",
-                    "emailaddress",
                     "fulltokenid",
                     "httpproxypassword",
                     "keypassword",
@@ -55,7 +50,6 @@ public final class SensitiveUtils {
                     "oauthclientsecret",
                     "oauthtoken",
                     "oauthtokenurl",
-                    "p12filename",
                     "passcode",
                     "passphrase",
                     "password",
@@ -68,12 +62,10 @@ public final class SensitiveUtils {
                     "proxypassword",
                     "proxyuser",
                     "publickeyid",
-                    "queueownerawsaccountid",
                     "refreshtoken",
                     "sasljaasconfig",
                     "secretkey",
                     "securerandom",
-                    "sharedaccesskey",
                     "sslkeypassword",
                     "sslkeystore",
                     "sslkeystorepassword",
